requirements.txt
nonebot==1.9.1
openai==2.7.1
Pillow==12.0.0
dashscope==1.24.9
